+++ src/components/SearchBar/SearchModal.js
@@ -1,15 +1,18 @@
-import React from 'react';
 import { createAutocomplete } from '@algolia/autocomplete-core';
+import React from 'react';
+import { useCurrentSection } from '../../hooks/useCurrentSection';
 import { createGetSources, navigator } from './autocomplete';
+import ResultsContainer from './ResultsContainer';
 import SearchForm from './SearchForm';
 import { useTouchEvents } from './useTouchEvents';
-import ResultsContainer from './ResultsContainer';
 
-const SearchModal = ({ searchClient, onClose }) => {
+const SearchModal = ({ searchClient, onClose, filterBySection, handleToggleFilterBySection }) => {
   // Refs to elements of search, to use with autocomplete-core
   const formElementRef = React.useRef(null);
   const inputRef = React.useRef(null);
   const resultsContainerRef = React.useRef(null);
+
+  const currentSection = useCurrentSection();
 
   // Store autocomplete's internal state on this component
   const [state, setState] = React.useState({
@@ -26,20 +29,27 @@ const SearchModal = ({ searchClient, onClose }) => {
     return createAutocomplete({
       id: 'resumos-search',
       defaultActiveItemId: 0,
-      placeholder: 'Search entire site...',
+      placeholder: `Search ${(filterBySection && currentSection) || 'entire site'}...`,
       openOnFocus: true,
       navigator,
       onStateChange(props) {
         setState(props.state);
       },
-      getSources: createGetSources({ searchClient, onClose }),
+      getSources: createGetSources({
+        searchClient,
+        onClose,
+        section: filterBySection && currentSection,
+      }),
     });
-  }, [searchClient]);
+  }, [searchClient, filterBySection, currentSection, onClose]);
 
-  const onItemClick = React.useCallback((item) => {
-    // In the future, we might want to save recent searches
-    onClose();
-  }, []);
+  const onItemClick = React.useCallback(
+    (item) => {
+      // In the future, we might want to save recent searches
+      onClose();
+    },
+    [onClose]
+  );
 
   const { getEnvironmentProps, getInputProps, getListProps, getItemProps } = autocomplete;
 
@@ -63,7 +73,26 @@ const SearchModal = ({ searchClient, onClose }) => {
           onItemClick={onItemClick}
         />
       </div>
-      <div className='search-footer'></div>
+      <div className='search-footer'>
+        {currentSection && (
+          <button className='search-filterbysection' onClick={handleToggleFilterBySection}>
+            <span
+              className={`search-filterbysection--btn ${
+                filterBySection ? 'search-filterbysection--btn__active' : ''
+              }`}
+            >
+              {`${currentSection} only`}
+            </span>
+            <span
+              className={`search-filterbysection--btn ${
+                filterBySection ? '' : 'search-filterbysection--btn__active'
+              }`}
+            >
+              Entire site
+            </span>
+          </button>
+        )}
+      </div>
     </>
   );
 };
